Fix: Buffer/SimpleBuffer.Substring> Length was shorter than expected.
string.Substring() requires length as 2nd argument. If I simply
subtract 'beg' from 'end', it is 1 shorter than what it should be.

Supporse (beg = 2, end = 5). As 'end' is inclusive in
IBuffer.Substring, it is expected to return string of [2, 3, 4, 5].

In previous implementation, it was [2, 3, 4] because length is
(5 - 2 = 3). So I simply added 1 to that length so that it will return
[2, 3, 4, 5] properly